refactoring for python3 compatibility
