+++ Dockerfile.test
@@ -3,13 +3,13 @@ FROM gettyimages/spark:2.1.0-hadoop-2.7
 RUN apt-get update && \
     DEBIAN_FRONTEND=noninteractive apt-get -yq install \
         libpq-dev \
-        python-pip \
-        python-dev \
+        python3-dev \
+        build-essential \
         && \
     rm -rf /var/lib/apt/lists/*
 
 ADD requirements-test.txt /tmp/
-RUN pip2 install -r /tmp/requirements-test.txt
+RUN pip3 install -r /tmp/requirements-test.txt
 RUN mkdir /srv/spark-stat-analyzer
 WORKDIR /srv/spark-stat-analyzer
