+++ analyzers/__init__.py
@@ -1,9 +1,9 @@
-from analyzer import Analyzer
-from error_stats import AnalyzeErrors
-from coverage_modes import AnalyzeCoverageModes
-from coverage_stop_areas import AnalyzeCoverageStopAreas
-from coverage_journeys_requests_params import AnalyzeCoverageJourneysRequestsParams
-from requests_calls import AnalyzeRequest
-from token_stats import AnalyzeTokens
-from users_sql import AnalyseUsersSql
-from coverage_journeys_transfers import AnalyzeCoverageJourneysTransfers
+from analyzers.analyzer import Analyzer
+from analyzers.error_stats import AnalyzeErrors
+from analyzers.coverage_modes import AnalyzeCoverageModes
+from analyzers.coverage_stop_areas import AnalyzeCoverageStopAreas
+from analyzers.coverage_journeys_requests_params import AnalyzeCoverageJourneysRequestsParams
+from analyzers.requests_calls import AnalyzeRequest
+from analyzers.token_stats import AnalyzeTokens
+from analyzers.users_sql import AnalyseUsersSql
+from analyzers.coverage_journeys_transfers import AnalyzeCoverageJourneysTransfers
