+++ analyzers/coverage_journeys_transfers.py
@@ -38,7 +38,7 @@ class AnalyzeCoverageJourneysTransfers(Analyzer):
             data = rdd.flatMap(
                 self.get_tuples_from_stat_dict
             ).reduceByKey(
-                lambda (a), (b): (a + b)
+                lambda a, b: (a + b)
             ).collect()
             return [tuple(list(key_tuple) + [nb]) for (key_tuple, nb) in data]
         else:
