+++ analyzers/coverage_modes.py
@@ -46,9 +46,9 @@ class AnalyzeCoverageModes(Analyzer):
                             datetime.utcfromtimestamp(stat_dict['request_date']).date()
                         )
                     ),
-                (
-                    1
-                )
+                    (
+                        1
+                    )
                 )
             )
         return result
@@ -58,7 +58,7 @@ class AnalyzeCoverageModes(Analyzer):
             coverage_modes = rdd.flatMap(
                 self.get_tuples_from_stat_dict
             ).reduceByKey(
-                lambda (a), (b): (a + b)
+                lambda a, b: (a + b)
             ).collect()
             return [tuple(list(key_tuple) + [nb]) for (key_tuple, nb) in coverage_modes]
         else:
