+++ tests/coverage_journeys_transfers_test.py
@@ -2,7 +2,7 @@ import pytest
 import os
 from datetime import date, datetime
 from analyzers import AnalyzeCoverageJourneysTransfers
-from checker import same_list_tuple
+from tests.checker import same_list_tuple
 
 pytestmark = pytest.mark.usefixtures("spark")
 path = os.getcwd() + "/tests/fixtures/coverage_journeys_transfers"
